pages for edit and add separately
