+++ event-calendar/app/components/form/SubmitButton.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@chakra-ui/react";
 import { useIsSubmitting } from "remix-validated-form";
 
-export const MySubmitButton = () => {
+export const MySubmitButton = ({ value }: { value: string }) => {
   const isSubmitting = useIsSubmitting();
   return (
     <Button
@@ -11,6 +11,8 @@ export const MySubmitButton = () => {
       textColor="white"
       _hover={{ bg: "white", textColor: "brand.200" }}
       marginTop={10}
+      name="action"
+      value={value}
     >
       {isSubmitting ? "Submitting..." : "Submit"}
     </Button>
