+++ event-calendar/app/routes/admin.categories.$categoryId.delete..tsx
